requirements.txt
streamlit
Pillow
e2b-code-interpreter
agno
